api/client_test.go
package api

import (
	"bytes"
	"database/sql"
	"encoding/json"
	"fmt"
	"io"
	"net/http"
	"net/http/httptest"
	mockdb "super-pet-delivery/db/mock"
	db "super-pet-delivery/db/sqlc"
	"super-pet-delivery/token"
	"super-pet-delivery/util"
	"testing"
	"time"

	"github.com/stretchr/testify/require"
	"go.uber.org/mock/gomock"
)

func TestCreateClientAPI(t *testing.T) {
	// Generate a random client for testing.
	client := randomClient()
	// Create a request body with client data.
	requestClient := createClientRequest{
		FullName:            client.FullName,
		PhoneWhatsapp:       client.PhoneWhatsapp,
		PhoneLine:           client.PhoneLine,
		PetName:             client.PetName,
		PetBreed:            client.PetBreed,
		AddressStreet:       client.AddressStreet,
		AddressNumber:       client.AddressNumber,
		AddressNeighborhood: client.AddressNeighborhood,
		AddressReference:    client.AddressReference,
	}

	testCases := []struct {
		name          string
		requestClient createClientRequest
		setupAuth     func(t *testing.T, request *http.Request, tokenMkaer token.Maker)
		buildStubs    func(store *mockdb.MockStore)
		checkResponse func(t *testing.T, recorder *httptest.ResponseRecorder)
	}{
		{
			name:          "OK",
			requestClient: requestClient,
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				// Define expectations for the CreateClient function in your mock store.
				store.EXPECT().CreateClient(gomock.Any(), gomock.Any()).Times(1).Return(client, nil)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusOK, recorder.Code)
				requireBodyMatchClient(t, recorder.Body, client)
			},
		},
		{
			name:          "BadRequest",
			requestClient: createClientRequest{}, // Sending an empty request should trigger a bad request.
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().CreateClient(gomock.Any(), gomock.Any()).Times(0) // Expect CreateClient not to be called.
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusBadRequest, recorder.Code)
			},
		},
		{
			name:          "InternalServerError",
			requestClient: requestClient,
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().CreateClient(gomock.Any(), gomock.Any()).Times(1).Return(db.Client{}, sql.ErrConnDone)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusInternalServerError, recorder.Code)
			},
		},
	}

	for i := range testCases {
		tc := testCases[i]

		t.Run(tc.name, func(t *testing.T) {
			ctrl := gomock.NewController(t)
			defer ctrl.Finish()

			store := mockdb.NewMockStore(ctrl)
			tc.buildStubs(store)

			server := newTestServer(t, store)
			recorder := httptest.NewRecorder()

			// Marshal the request body into JSON.
			requestBody, err := json.Marshal(tc.requestClient)
			require.NoError(t, err)

			// Create an HTTP request with the JSON body.
			request, err := http.NewRequest(http.MethodPost, "/clients", bytes.NewReader(requestBody))
			require.NoError(t, err)

			// Serve the request and check the response.
			tc.setupAuth(t, request, server.tokenMaker)
			server.router.ServeHTTP(recorder, request)
			tc.checkResponse(t, recorder)
		})
	}
}

// TestGetClientAPI tests the GetClient API endpoint.
func TestGetClientAPI(t *testing.T) {
	// Generate a random client for testing.
	client := randomClient()

	testCases := []struct {
		name          string
		clientID      int64
		setupAuth     func(t *testing.T, request *http.Request, tokenMaker token.Maker)
		buildStubs    func(store *mockdb.MockStore)
		checkResponse func(t *testing.T, recorder *httptest.ResponseRecorder)
	}{
		{
			name:     "OK",
			clientID: client.ID,
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().GetClient(gomock.Any(), gomock.Eq(client.ID)).Times(1).Return(client, nil)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusOK, recorder.Code)
				requireBodyMatchClient(t, recorder.Body, client)
			},
		},
		{
			name:     "NotFound",
			clientID: client.ID,
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().GetClient(gomock.Any(), gomock.Eq(client.ID)).Times(1).Return(db.Client{}, sql.ErrNoRows)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusNotFound, recorder.Code)
			},
		},
		{
			name:     "InternalError",
			clientID: client.ID,
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().GetClient(gomock.Any(), gomock.Eq(client.ID)).Times(1).Return(db.Client{}, sql.ErrConnDone)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusInternalServerError, recorder.Code)
			},
		},
		{
			name:     "InvalidID",
			clientID: 0,
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().GetClient(gomock.Any(), gomock.Any()).Times(0)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusBadRequest, recorder.Code)
			},
		},
		// TODO: add more cases
	}

	for i := range testCases {
		tc := testCases[i]

		t.Run(tc.name, func(t *testing.T) {

			ctrl := gomock.NewController(t)
			defer ctrl.Finish()

			store := mockdb.NewMockStore(ctrl)
			tc.buildStubs(store)

			server := newTestServer(t, store)
			recorder := httptest.NewRecorder()

			// Create an HTTP request with the client ID.
			url := fmt.Sprintf("/clients/%d", tc.clientID)
			request, err := http.NewRequest(http.MethodGet, url, nil)
			require.NoError(t, err)

			// Use the setupAuth function to add authentication to the request.
			tc.setupAuth(t, request, server.tokenMaker)

			// Serve the request and check the response.
			server.router.ServeHTTP(recorder, request)
			tc.checkResponse(t, recorder)
		})

	}
}

// TestListClientAPI tests the ListClients API endpoint.
func TestListClientAPI(t *testing.T) {
	n := 5
	clients := make([]db.Client, n)
	for i := 0; i < n; i++ {
		clients[i] = randomClient()
	}

	type Query struct {
		pageID   int
		pageSize int
	}

	testCases := []struct {
		name          string
		query         Query
		setupAuth     func(t *testing.T, request *http.Request, tokenMaker token.Maker)
		buildStubs    func(store *mockdb.MockStore)
		checkResponse func(recorder *httptest.ResponseRecorder)
	}{
		{
			name: "OK",
			query: Query{
				pageID:   1,
				pageSize: n,
			},
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				arg := db.ListClientsParams{
					Limit:  int32(n),
					Offset: 0,
				}

				store.EXPECT().
					ListClients(gomock.Any(), gomock.Eq(arg)).
					Times(1).
					Return(clients, nil)
			},
			checkResponse: func(recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusOK, recorder.Code)
				requireBodyMatchClients(t, recorder.Body, clients)
			},
		},
		{
			name: "InternalError",
			query: Query{
				pageID:   1,
				pageSize: n,
			},
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().
					ListClients(gomock.Any(), gomock.Any()).
					Times(1).
					Return([]db.Client{}, sql.ErrConnDone)
			},
			checkResponse: func(recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusInternalServerError, recorder.Code)
			},
		},
		{
			name: "InvalidPageID",
			query: Query{
				pageID:   -1,
				pageSize: n,
			},
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().
					ListClients(gomock.Any(), gomock.Any()).
					Times(0)
			},
			checkResponse: func(recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusBadRequest, recorder.Code)
			},
		},
		{
			name: "InvalidPageSize",
			query: Query{
				pageID:   1,
				pageSize: 100000,
			},
			setupAuth: func(t *testing.T, request *http.Request, tokenMaker token.Maker) {
				addAuthorization(t, request, tokenMaker, authorizationTypeBearer, "username", time.Minute)
			},
			buildStubs: func(store *mockdb.MockStore) {
				store.EXPECT().
					ListClients(gomock.Any(), gomock.Any()).
					Times(0)
			},
			checkResponse: func(recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusBadRequest, recorder.Code)
			},
		},
	}

	for i := range testCases {
		tc := testCases[i]

		t.Run(tc.name, func(t *testing.T) {
			ctrl := gomock.NewController(t)
			defer ctrl.Finish()

			store := mockdb.NewMockStore(ctrl)
			tc.buildStubs(store)

			server := newTestServer(t, store)
			recorder := httptest.NewRecorder()

			url := "/clients"
			request, err := http.NewRequest(http.MethodGet, url, nil)
			require.NoError(t, err)

			// Add query parameters to request URL.
			q := request.URL.Query()
			q.Add("page_id", fmt.Sprintf("%d", tc.query.pageID))
			q.Add("page_size", fmt.Sprintf("%d", tc.query.pageSize))
			request.URL.RawQuery = q.Encode()

			// Use the setupAuth function to add authentication to the request.
			tc.setupAuth(t, request, server.tokenMaker)

			// Serve the request and check the response.
			server.router.ServeHTTP(recorder, request)
			tc.checkResponse(recorder)
		})
	}
}

// TODO fix here not passing
/* func TestUpdateClientAPI(t *testing.T) {
	// Generate a random client for testing.
	client := randomClient()
	// Create a request body with client data.
	updateClient := updateClientRequest{
		FullName: "New Full Name",
	}

	testCases := []struct {
		name          string
		userID        int64
		requestClient updateClientRequest
		buildStubs    func(store *mockdb.MockStore)
		checkResponse func(t *testing.T, recorder *httptest.ResponseRecorder)
	}{
		{
			name:          "OK",
			userID:        client.ID,
			requestClient: updateClient,
			buildStubs: func(store *mockdb.MockStore) {
				// Define expectations for the UpdateClient function in your mock store.
				arg := db.UpdateClientParams{
					ID:       client.ID,
					FullName: updateClient.FullName,
				}
				store.EXPECT().GetClient(gomock.Any(), gomock.Eq(client.ID)).Times(1).Return(client, nil)
				store.EXPECT().UpdateClient(gomock.Any(), gomock.Eq(arg)).Times(1).Return(client, nil)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusOK, recorder.Code)
				requireBodyMatchClient(t, recorder.Body, client)
			},
		},
	}

	for i := range testCases {
		tc := testCases[i]

		t.Run(tc.name, func(t *testing.T) {
			ctrl := gomock.NewController(t)
			defer ctrl.Finish()

			store := mockdb.NewMockStore(ctrl)
			tc.buildStubs(store)

			server := newTestServer(t, store)
			recorder := httptest.NewRecorder()

			// Marshal the request body into JSON.
			requestBody, err := json.Marshal(tc.requestClient)
			require.NoError(t, err)

			// Create an HTTP request with the JSON body.
			url := fmt.Sprintf("/clients/%d", tc.userID)
			request, err := http.NewRequest(http.MethodPut, url, bytes.NewReader(requestBody))
			require.NoError(t, err)

			// Serve the request and check the response.
			server.router.ServeHTTP(recorder, request)
			tc.checkResponse(t, recorder)
		})
	}
} */

// TODO FIX HERE NOT PASSING
/* func TestDeleteClientAPI(t *testing.T) {
	// Generate a random client for testing.
	client := randomClient()

	testCases := []struct {
		name          string
		clientID      int64
		request       deleteClientRequest
		buildStubs    func(store *mockdb.MockStore)
		checkResponse func(t *testing.T, recorder *httptest.ResponseRecorder)
	}{
		{
			name:     "OK",
			clientID: client.ID,
			buildStubs: func(store *mockdb.MockStore) {
				// Expect GetClient to return the client.
				store.EXPECT().GetClient(gomock.Any(), gomock.Eq(client.ID)).Times(1).Return(client, nil)
				// Expect DeleteClient to delete the client.
				store.EXPECT().DeleteClient(gomock.Any(), gomock.Eq(client.ID)).Times(1).Return(nil)
			},
			checkResponse: func(t *testing.T, recorder *httptest.ResponseRecorder) {
				require.Equal(t, http.StatusOK, recorder.Code)
				// Verify the response message.
				responseBody, err := io.ReadAll(recorder.Body)
				require.NoError(t, err)
				// Updated expected response format to match the actual response.
				require.Equal(t, `"Client deleted successfully"`, string(responseBody))
			},
		},
	}

	for i := range testCases {
		tc := testCases[i]

		t.Run(tc.name, func(t *testing.T) {
			ctrl := gomock.NewController(t)
			defer ctrl.Finish()

			store := mockdb.NewMockStore(ctrl)
			tc.buildStubs(store)

			server := newTestServer(t, store)
			recorder := httptest.NewRecorder()

			// Marshal the request body into JSON.
			requestBody, err := json.Marshal(tc.request)
			require.NoError(t, err)

			// Create an HTTP request with the JSON body.
			url := fmt.Sprintf("/clients/%d", tc.clientID)
			request, err := http.NewRequest(http.MethodDelete, url, bytes.NewReader(requestBody))
			require.NoError(t, err)

			// Serve the request and check the response.
			server.router.ServeHTTP(recorder, request)
			tc.checkResponse(t, recorder)
		})
	}
} */

// randomClient generates a random client for testing.
func randomClient() db.Client {
	return db.Client{
		ID:                  util.RandomInt(1, 1000),
		FullName:            util.RandomFullName(),
		PhoneWhatsapp:       util.RandomString(9),
		PhoneLine:           util.RandomString(9),
		PetName:             util.RandomString(9),
		PetBreed:            util.RandomString(9),
		AddressStreet:       util.RandomString(9),
		AddressNumber:       util.RandomString(9),
		AddressNeighborhood: util.RandomString(9),
		AddressReference:    util.RandomString(9),
	}
}

// requireBodyMatchClient checks if the response body matches the expected client.
func requireBodyMatchClient(t *testing.T, body *bytes.Buffer, client db.Client) {
	data, err := io.ReadAll(body)
	require.NoError(t, err)

	var gotClient db.Client
	err = json.Unmarshal(data, &gotClient)
	require.NoError(t, err)
	require.Equal(t, client, gotClient)
}

// requireBodyMatchClients checks if the response body matches the expected list of clients.
func requireBodyMatchClients(t *testing.T, body *bytes.Buffer, clients []db.Client) {
	data, err := io.ReadAll(body)
	require.NoError(t, err)

	var gotClients []db.Client
	err = json.Unmarshal(data, &gotClients)
	require.NoError(t, err)
	require.Equal(t, clients, gotClients)
}
